+++ packages/ember-input/src/test-support/datepicker-helpers.ts
@@ -1,4 +1,4 @@
-import type { HTMLInputTDElement } from '../components/tpk-datepicker/input';
+import type { HTMLInputTDElement } from '../components/tpk-datepicker-input';
 import { find } from '@ember/test-helpers';
 import { DateTime } from '@eonasdan/tempus-dominus';
